pyproject.toml
[tool.poetry]
name = "python-wm"
version = "0.1.0"
description = ""
authors = ["Wilkne Maia <72803395+wilkneMaia@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.1"
camelot-py = "^0.11.0"
opencv-python = "^4.10.0.84"
matplotlib = "^3.9.2"
ghostscript = "^0.7"
pandas = "^2.2.2"
unidecode = "^1.3.8"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
